almost finished cart crud

+++ app/forms/order_form.py
@@ -1,8 +1,8 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, TextAreaField, IntegerField
+from wtforms import StringField, TextAreaField, DecimalField
 from wtforms.validators import DataRequired
 
 class OrderForm(FlaskForm):
     address = StringField("address", validators=[DataRequired()])
-    price = IntegerField("price", validators=[DataRequired()])
+    price = DecimalField("price", validators=[DataRequired()])
